Add new pfe_run tool
The pfe_run tool will reset and initialize the PFE to get a working Rx
path using the HIF_NOCPY.
This does not require loading any firmware blob. Actually, the PE cores
are held in reset.
To be able to receive packets, the ethernet PHYs must still be powered
and initialized.